requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-dotenv==1.1.1
pandas==2.3.3
httpx==0.28.1
tensorflow==2.20.0
scikit_learn==1.7.2